src/requirements.txt
asgiref==3.5.2
celery==5.2
Django==4.1.8
django-filter==2.4.0
djangorestframework==3.13.1
drf-yasg==1.21.3
flower==1.2.0
gunicorn==20.0.4
# psycopg2-binary==2.9.5
psycopg2==2.9.3
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.3
sqlparse==0.3.1
requests==2.26.0
redis==4.5.4
python-dotenv==1.0.0
flake8==5.0.4
flake8-isort==6.0.0
flake8-return==1.2.0
flake8-broken-line==0.6.0
pep8-naming==0.13.3
pytest-cov==4.1.0